// app/CasterWeb/src/api/types.ts
export enum SourceRecordType {
  SOURCE_RECORD_TYPE_UNKNOWN = 0,
  SOURCE_RECORD_TYPE_REAL = 1,
  SOURCE_RECORD_TYPE_NEAREST = 2,
  SOURCE_RECORD_TYPE_ALIAS = 3,
  SOURCE_RECORD_TYPE_PROXY = 4,
  SOURCE_RECORD_TYPE_GRID = 5,
}

export enum SourceDecordType {
  SOURCE_DECODE_TYPE_UNKNOWN = 0,
  SOURCE_DECODE_TYPE_AUTO = 1,
  SOURCE_DECODE_TYPE_MANUAL = 2,
}

export enum SourceDisplayType {
  SOURCE_DISP_TYPE_UNKNOWN = 0,
  SOURCE_DISP_TYPE_ALWAYS_SHOW = 1,
  SOURCE_DISP_TYPE_ALWAYS_HIDE = 2,
  SOURCE_DISP_TYPE_SHOW_WHEN_ONLINE = 3,
}
// ...
export interface SourceRecord {
  uid: string;
  create_time: number;
  update_time: number;
  source_group_uid: string;
  mountpoint: string;
  identufier: string;
  format: string;
  format_details: string;
  carrier: string;
  nav_system: string;
  network: string;
  country: string;
  latitude: string;
  longitude: string;
  nmea: string;
  solution: string;
  generator: string;
  compr_encrryp: string;
  authentication: string;
  fee: string;
  bitrate: string;
  misc: string;
  decode_type: SourceDecordType;
  display_type: SourceDisplayType;
  record_type: SourceRecordType;
}